mescc: Mes C Library: Fix compile warnings.
* lib/stdio/fputc.c (fputc): Oops, stream is a long.
* lib/stdlib/malloc.c (malloc): Cast to char *. FIXME
* lib/string/memchr.c (memchr): Cast to void *.
* lib/string/memcmp.c (memcmp): Add const cast.

+++ lib/stdio/fputc.c
@@ -23,5 +23,5 @@
 int
 fputc (int c, FILE * stream)
 {
-  return fdputc (c, (int) stream);
+  return fdputc (c, (long) stream);
 }
